2단계 피드백: userInfo parsing 유틸함수 작성
